requirements.txt
discord.py
flask
google-auth-oauthlib
google-api-python-client
